+++ lib/src/data_frame/data_frame_helpers.dart
@@ -1,4 +1,5 @@
 import 'package:ml_dataframe/src/data_frame/series.dart';
+import 'package:quiver/iterables.dart';
 
 Iterable<Series> convertRowsToSeries(
     Iterable<String> columnHeaders,
@@ -25,4 +26,24 @@ Iterable<Iterable<dynamic>> convertSeriesToRows(Iterable<Series> series) sync* {
   while (iterators.fold(true, (isActive, iterator) => iterator.moveNext())) {
     yield iterators.map<dynamic>((iterator) => iterator.current);
   }
+}
+
+Iterable<String> getHeader(
+    int columnsNum,
+    String autoHeaderPrefix,
+    [
+      Iterable<String> rawActualHeader = const [],
+      Iterable<String> predefinedHeader = const [],
+    ]
+) {
+  if (predefinedHeader?.isNotEmpty == true) {
+    return predefinedHeader.take(columnsNum);
+  }
+
+  if (rawActualHeader?.isNotEmpty == true) {
+    return rawActualHeader.map((name) => name.trim());
+  }
+
+  return count(0).take(columnsNum)
+      .map((index) => '${autoHeaderPrefix}${index}');
 }